Fix duplicate ids issue in post layout html
The post layout had 3 ids that were duplicated on pages with more than
one post: category, post-date and tags.

Each id was used to match a specific div element in an Enlive snippet.
Since the snippet of a post only contains the HTML of that post, a
category can be used instead of an id.